Add an option to use aya-bpf-cty instead of cty
We would like to be able to use cstr_core in eBPF programs based on
Aya[0]. However, cty crate, which cstr_core depends on, doesn't support
BPF arch targets (bpfel-unknown-none, bpfeb-unknown-none). aya-bpf-cty
is a crate which supports those targets.

This change adds the aya-bpf-cty feature which allows to use aya-bpf-cty
instead of cty.

[0] https://github.com/aya-rs/aya

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>